+++ src/Database/Beam/Schema.hs
@@ -7,7 +7,7 @@
 module Database.Beam.Schema
     ( module Database.Beam.Schema.Tables
     , module Database.Beam.Schema.Fields
-    , module Database.Beam.Schema.Lenses ) where
+    , module Database.Beam.Schema.Lenses) where
 
 import Database.Beam.Schema.Tables
 import Database.Beam.Schema.Fields